Added enemy health bar base.  Dragonfly and Raptor offsets are complete, rest of ships are still off.

+++ Save Earth/Assets/Scripts/Controllers/Ship.cs
@@ -11,7 +11,7 @@ public class Ship : MonoBehaviour {
  
 
 	[HideInInspector]
-    public int health;
+    public float health;
 
 
     #region Movement Methods
